Inline FullConnectorSecurityContext
