+++ presto-spi/src/main/java/io/prestosql/spi/connector/ConnectorSecurityContext.java
@@ -15,9 +15,26 @@ package io.prestosql.spi.connector;
 
 import io.prestosql.spi.security.ConnectorIdentity;
 
-public interface ConnectorSecurityContext
-{
-    ConnectorTransactionHandle getTransactionHandle();
+import static java.util.Objects.requireNonNull;
 
-    ConnectorIdentity getIdentity();
+public class ConnectorSecurityContext
+{
+    private final ConnectorTransactionHandle transactionHandle;
+    private final ConnectorIdentity identity;
+
+    public ConnectorSecurityContext(ConnectorTransactionHandle transactionHandle, ConnectorIdentity identity)
+    {
+        this.transactionHandle = requireNonNull(transactionHandle, "transactionHandle is null");
+        this.identity = requireNonNull(identity, "identity is null");
+    }
+
+    public ConnectorTransactionHandle getTransactionHandle()
+    {
+        return transactionHandle;
+    }
+
+    public ConnectorIdentity getIdentity()
+    {
+        return identity;
+    }
 }
